Fix beta application status for organizations
